provider/marathon: pass cmd via env
This PR reworks entrypoint to execute cmd read from environment
variable. This fixes an issue, when global command was used for
all of the containers, causing only one entrypoint to be executed.
Instead the global command is removed and passed as env, so each
container executes a unique entrypoint, which then executes cmd.