Bongo: invokeMethod: fix all the unhandled calls issues
This is very annoying that I just realised all the calls drops here if
there is an error in the invoker. For example, if you make an instance
call on a non-existent instance it never calls the callback. Or if you
try to call a non-existent method over any Bongo.model this never calls
the callback as well. It was valid for lots of other cases as well in
this commit.

With this commit we will call the provided callback with a proper error
instead. which I assume will fix the remaining memory issues on Bongo.